Updated identifier on DELETE /site/<id> to SaaS for Issue #2

+++ includes/class-trustedlogin.php
@@ -1133,8 +1133,15 @@ class TrustedLogin
         );
 
         if ('revoke' == $action) {
+
+            $keys = get_site_option('tl_' . $this->ns . '_slt', false);
+
+            if (!$keys || !is_array($keys) || !array_key_exists('authToken', $keys)) {
+                $this->dlog('Delete key not found in local db: ' . print_r($keys, true), __METHOD__);
+                return false;
+            }
             $method = 'DELETE';
-            $endpoint = 'sites/' . $vault_id;
+            $endpoint = 'sites/' . $keys['authToken'];
         } else {
             $method = 'POST';
             $endpoint = 'sites';
